// client/src/state/BuildingState.ts
export const getBuildingsAsync = createAsyncThunk(
    'getBuildingsAsync',
    async () => {
        const getBuildingsUrl = BackendBaseUrl + "/building/all"
        const data = await axios.get(
            getBuildingsUrl,
            {
                headers: {
                    token: window.localStorage.getItem(Keys.Access_token)
                },
                withCredentials: true
            }
        )
        console.log(data)
        return {buildings: data.data.buildings}
    }
);

export const createBuildingAsync = createAsyncThunk(
    'createBuildingAsync',
    async ({name, address}:{name: string, address: string}) => {
        const createBuildingAsyncUrl = BackendBaseUrl + "/building/create"
        const data = await axios.post(
            createBuildingAsyncUrl,
            {
                name : name,
                address : address
            },
            {
                headers: {
                    token: window.localStorage.getItem(Keys.Access_token)
                },
                withCredentials: true
            }
        )
        console.log(data)
        return data
    }
);

export const deleteBuildingAsync = createAsyncThunk(
    'deleteBuildingAsync',
    async (buildingId: string) => {
        const deleteBuildingAsyncUrl = BackendBaseUrl + "/building/delete"
        console.log(buildingId)
        const data = await axios.post(
            deleteBuildingAsyncUrl,
            {
                buildingId : buildingId
            },
            {
                headers: {
                    token: window.localStorage.getItem(Keys.Access_token)
                },
                withCredentials: true
            }
        )
        console.log(data)
        return data
    }
);