// client/src/state/RoomState.ts
export const createRoomAsync = createAsyncThunk(
    'createRoomAsync',
    async ({name, floor, buildingId}:{name: string, floor: number, buildingId: string}) => {
        const createRoomAsyncUrl = BackendBaseUrl + "/room/create"
        const data = await axios.post(
            createRoomAsyncUrl,
            {
                name : name,
                floor : floor,
                buildingId: buildingId
            },
            {
                headers: {
                    token: window.localStorage.getItem(Keys.Access_token)
                },
                withCredentials: true
            }
        )
        console.log(data)
        return data
    }
);

export const deleteRoomAsync = createAsyncThunk(
    'deleteRoomAsync',
    async (roomId: string) => {
        const deleteRoomAsyncUrl = BackendBaseUrl + "/room/delete"
        const data = await axios.post(
            deleteRoomAsyncUrl,
            {
                roomId : roomId
            },
            {
                headers: {
                    token: window.localStorage.getItem(Keys.Access_token)
                },
                withCredentials: true
            }
        )
        console.log(data)
        return data
    }
);